update waiting specs
