+++ spec/watirspec/support/rspec_matchers.rb
@@ -117,27 +117,28 @@ if defined?(RSpec)
     end
   end
 
-  RSpec::Matchers.define :execute_immediately do |timeout: 10|
+  RSpec::Matchers.define :execute_when_satisfied do |min: 0, max: nil|
+    max ||= min + 1
     match do |actual|
       original_timeout = Watir.default_timeout
-      Watir.default_timeout = timeout
+      Watir.default_timeout = max
       begin
         start_time = ::Time.now
         actual.call
         @time_difference = ::Time.now - start_time
-        @time_difference < timeout
+        @time_difference > min && @time_difference < max
       ensure
         Watir.default_timeout = original_timeout
       end
     end
 
     failure_message_when_negated do
-      "expected action to take more than provided timeout (#{timeout} seconds), " \
+      "expected action to take less than #{min} seconds or more than #{max} seconds, " \
       "instead it took #{@time_difference} seconds"
     end
 
     failure_message do
-      "expected action to take less than provided timeout (#{timeout} seconds), " \
+      "expected action to take more than #{min} seconds and less than #{max} seconds, " \
       "instead it took #{@time_difference} seconds"
     end
